+++ src/app/components/users-groups/users-groups.component.ts
@@ -2,6 +2,9 @@ import {Component, OnInit} from '@angular/core';
 import {Group} from '../../models/Group';
 import {GroupServiceService} from '../../services/group-service/group-service.service';
 import {Router} from '@angular/router';
+import {MatDialog} from '@angular/material/dialog';
+import {DeleteDialogComponent} from '../dialogs/delete-dialog/delete-dialog.component';
+import {RestApiServiceService} from '../../services/restapiservice/rest-api-service.service';
 
 @Component({
   selector: 'app-users-groups',
@@ -10,7 +13,9 @@ import {Router} from '@angular/router';
 })
 export class UsersGroupsComponent implements OnInit {
 
-  constructor(private groupService: GroupServiceService, private router: Router) {
+  constructor(private groupService: GroupServiceService,
+              private router: Router,
+              public dialog: MatDialog) {
   }
 
   Groups: Group[];
@@ -25,8 +30,14 @@ export class UsersGroupsComponent implements OnInit {
   }
 
   deleteGroup(group) {
-    this.groupService.deleteGroupByName(group).subscribe(data => {
-      this.getAllGroups();
+    let dialogRef = this.dialog.open(DeleteDialogComponent);
+
+    dialogRef.afterClosed().subscribe(data => {
+      if (data == 'true') {
+        this.groupService.deleteGroupByName(group).subscribe(data => {
+          this.getAllGroups();
+        });
+      }
     });
   }
